feat: moved core faker implementation to mpp jvmMain and dropped legacy module
