--- core-mpp/build.gradle.kts
@@ -1,42 +1,0 @@
-plugins {
-  kotlin("multiplatform")
-  id("io.bkbn.sourdough.library.mpp") version "0.5.6"
-  id("io.gitlab.arturbosch.detekt") version "1.19.0"
-  id("com.adarshr.test-logger") version "3.1.0"
-  id("org.jetbrains.dokka")
-//  id("maven-publish")
-}
-
-sourdough {
-  githubOrg.set("unredundant")
-  githubRepo.set("satisfaketion")
-  libraryName.set("Satisfaketion")
-  libraryDescription.set("A data generator that is as beautiful and powerful as you are ❤️")
-  licenseName.set("MIT License")
-  licenseUrl.set("https://mit-license.org")
-  developerId.set("unredundant")
-  developerName.set("Ryan Brink")
-  developerEmail.set("admin@bkbn.io")
-}
-
-dependencies {
-  implementation(group = "org.jetbrains.kotlin", name = "kotlin-reflect", version = "1.6.10")
-  detektPlugins(group = "io.gitlab.arturbosch.detekt", name = "detekt-formatting", version = "1.19.0")
-}
-
-kotlin {
-  sourceSets {
-    val commonMain by getting
-    val commonTest by getting {
-      dependencies {
-        implementation(kotlin("test"))
-      }
-    }
-    val jvmMain by getting
-    val jvmTest by getting
-    val jsMain by getting
-    val jsTest by getting
-    val nativeMain by getting
-    val nativeTest by getting
-  }
-}
