--- core-mpp/src/commonMain/kotlin/io/github/unredundant/satisfaketion/core/Faker.kt
@@ -1,7 +1,0 @@
-package io.github.unredundant.satisfaketion.core
-
-import kotlin.random.Random
-
-interface Faker<T> {
-  fun generate(seed: Random): T
-}
