--- core-mpp/src/commonMain/kotlin/io/github/unredundant/satisfaketion/core/FakerFactory.kt
@@ -1,5 +1,0 @@
-package io.github.unredundant.satisfaketion.core
-
-expect object FakerFactory {
-  fun <T> createFaker(): Faker<T>
-}
