--- core-mpp/src/nativeMain/kotlin/io/github/unredundant/satisfaketion/core/FakerFactory.kt
@@ -1,7 +1,0 @@
-package io.github.unredundant.satisfaketion.core
-
-actual object FakerFactory {
-  actual fun <T> createFaker(): Faker<T> {
-    TODO("Not yet implemented")
-  }
-}
