+++ core/src/commonMain/kotlin/io/github/unredundant/satisfaketion/core/Extensions.kt
