+++ core/src/commonMain/kotlin/io/github/unredundant/satisfaketion/core/Faker.kt
@@ -1,0 +1,8 @@
+package io.github.unredundant.satisfaketion.core
+
+import kotlin.random.Random
+
+expect class Faker<T : Any> {
+  fun generate(): T
+  fun generate(seed: Random): T
+}
