+++ core/src/commonMain/kotlin/io/github/unredundant/satisfaketion/core/Mutator.kt
