+++ core/src/jvmTest/kotlin/io/github/unredundant/satisfaketion/core/util/TestModels.kt
