--- core/src/main/kotlin/io/github/unredundant/satisfaketion/core/Mutator.kt
@@ -1,5 +1,0 @@
-package io.github.unredundant.satisfaketion.core
-
-fun interface Mutator<R, RR> {
-  fun mutate(generator: Generator<R>): Generator<RR>
-}
