+++ gradle.properties
@@ -1,5 +1,5 @@
 # Satisfaketion
-project.version=0.6.4
+project.version=0.7.0
 # Kotlin
 kotlin.code.style=official
 # Gradle
